refactor: change page path of courses list by region for RESTful URL
