--- app/courses/by-region/page.tsx
@@ -1,5 +1,0 @@
-import ListCourse from '@/src/views/list-course'
-
-export default function Page() {
-  return <ListCourse />
-}
